some bug fixes, a few swathes of code commented out some reorganising of the pokemon class

+++ src/main/scala/class_combat.scala
@@ -24,7 +24,7 @@ class Fight (team1:Team, team2:Team) {
         var fail_attack = true
         var nb_attack:Int = 0
         var att:Attack = new Attack("")
-        var bonus_typ:Int = 0
+        var bonus_typ:Double = 0.0
 
         var typ1:String = current_pok_ally.typ
         var typ2:String = current_pok_enemy.typ
@@ -40,7 +40,7 @@ class Fight (team1:Team, team2:Team) {
             att = current_pok_enemy.set_attack(nb_attack)
             fail_attack = att.use_attack()
             if (!(fail_attack)) { 
-                current_pok_ally.loss_PV(att.damage*(current_pok_ally.statDef+bonus_typ)*current_pok_enemy.statAtt)
+                current_pok_ally.loss_PV((att.damage*(current_pok_ally.statDef+bonus_typ)*current_pok_enemy.statAtt).toInt)
             }
         }
         nb_attack
